+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# Revision history for blocks
+# Revision history for `control-block`
+
+## 0.0.1 -- Unreleased
+
+* `perhaps` as a rearrangement of `maybe`
+* `plenty` to split a list into `[]` and `NonEmpty` cases
 
 ## 0.0.0 -- 2023-04-11